# src/main/resources/swagger.yml
openapi: 3.0.3
info:
  title: PlayBall API
  description: |
    **PlayBall** - 스포츠 매칭 플랫폼 REST API 명세서

    ## 인증 방식
    카카오 OAuth 로그인 후 서버에서 JWT를 발급합니다.
    - `POST /api/auth/kakao` → 카카오 인가코드로 로그인 → accessToken + refreshToken 발급
    - 인증 필요 API: `Authorization: Bearer {accessToken}` 헤더 포함
    - 비로그인 허용 API: `security: []` 표기

    ## 공통 응답
    ```json
    { "success": true, "message": "성공", "data": { ... } }
    ```

    ## 에러 응답
    ```json
    { "success": false, "message": "에러 메시지", "data": null }
    ```

    ## 종목 코드
    `SOCCER` | `BASKETBALL` | `RUNNING` | `BADMINTON`

    ## 실력 코드
    `BEGINNER` | `INTERMEDIATE` | `ADVANCED`
  version: 1.1.0
  contact:
    name: PlayBall Team
servers:
  - url: http://localhost:8080
    description: 로컬 개발
  - url: https://api.playball.kr
    description: 운영
...
security:
  - BearerAuth: []

# ============================================================
paths:
...
  /api/members/signup/complete:
    post:
      tags: [회원]
      summary: 회원가입 추가정보 입력
      description: |
        카카오 로그인 후 신규 회원(`isNewUser: true`)이 추가 정보를 입력합니다.
        이미 완료된 계정에서 재호출하면 409 에러를 반환합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpCompleteRequest'
            example:
              nickname: "홍길동"
              gender: "M"
              age: 22
              address: "서울시 노원구 신부동"
              latitude: 37.6394
              longitude: 127.0722
              favoriteSports: ["SOCCER", "BASKETBALL"]
              skillLevel: "INTERMEDIATE"
              preferredPosition: "ST"
      responses:
        '200':
          description: 가입 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: "회원가입이 완료되었습니다"
                data:
                  memberId: 2
                  nickname: "홍길동"
                  email: "hong@kakao.com"
                  gender: "M"
                  age: 22
                  address: "서울시 노원구 신부동"
                  favoriteSports: ["SOCCER", "BASKETBALL"]
                  skillLevel: "INTERMEDIATE"
                  preferredPosition: "ST"
                  role: "USER"
                  signupCompleted: true
        '409':
          description: 닉네임 중복 또는 이미 가입 완료된 계정
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                nicknameDuplicate:
                  summary: 닉네임 중복
                  value: { success: false, message: "이미 사용 중인 닉네임입니다.", data: null }
                alreadyCompleted:
                  summary: 이미 가입 완료
                  value: { success: false, message: "이미 회원가입이 완료된 계정입니다.", data: null }
  /api/members/check-nickname:
    get:
      tags: [회원]
      summary: 닉네임 중복 확인
      description: 비로그인 상태에서도 사용 가능합니다.
      security: []
      parameters:
        - name: nickname
          in: query
          required: true
          schema:
            type: string
          example: "홍길동"
      responses:
        '200':
          description: 확인 완료 (사용 가능 여부는 data.available 참고)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                available:
                  summary: 사용 가능
                  value: { success: true, message: "사용 가능한 닉네임입니다", data: { available: true } }
                duplicate:
                  summary: 중복
                  value: { success: true, message: "이미 사용 중인 닉네임입니다", data: { available: false } }
  /api/members/me:
    get:
      tags: [회원]
      summary: 내 프로필 조회
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: "프로필 조회 성공"
                data:
                  memberId: 1
                  nickname: "홍길동"
                  profileImage: "https://k.kakaocdn.net/..."
                  gender: "M"
                  age: 22
                  address: "서울시 노원구 신부동"
                  skillLevel: "INTERMEDIATE"
                  preferredPosition: "ST"
                  favoriteSports: ["SOCCER", "BASKETBALL"]
        '401':
          $ref: '#/components/responses/Unauthorized'

    patch:
      tags: [회원]
      summary: 내 프로필 수정
      description: |
        변경할 필드만 포함하면 됩니다. null 전달 시 해당 필드는 변경되지 않습니다.
        - `nickname`: null이면 유지, 빈 문자열("")은 검증 실패(2자 미만)
        - `favoriteSports`: null이면 유지, 빈 배열([])이면 전체 삭제
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
            example:
              nickname: "새닉네임"
              address: "서울시 강남구"
              latitude: 37.5665
              longitude: 126.9780
              favoriteSports: ["SOCCER", "BADMINTON"]
      responses:
        '200':
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: "프로필이 수정되었습니다"
                data:
                  memberId: 1
                  nickname: "새닉네임"
                  profileImage: "https://k.kakaocdn.net/..."
                  gender: "M"
                  age: 22
                  address: "서울시 강남구"
                  skillLevel: "INTERMEDIATE"
                  preferredPosition: "ST"
                  favoriteSports: ["SOCCER", "BADMINTON"]
        '409':
          description: 닉네임 중복
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example: { success: false, message: "이미 사용 중인 닉네임입니다.", data: null }
  /api/members/me/location:
    patch:
      tags: [회원]
      summary: 현재 위치 업데이트
      description: 지역 매칭 시 사용할 현재 위치를 갱신합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [latitude, longitude]
              properties:
                latitude:
                  type: number
                  format: double
                longitude:
                  type: number
                  format: double
                address:
                  type: string
            example:
              latitude: 37.6394
              longitude: 127.0722
              address: "서울시 노원구 신부동"
      responses:
        '200':
          description: 업데이트 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: "위치가 업데이트되었습니다"
                data: null
  /api/members/me/matches:
    get:
      tags: [회원]
      summary: 내가 참가한 경기 목록 조회
      description: |
        내가 APPROVED 상태로 참가한 경기 목록을 경기일 역순으로 반환합니다.
        DELETED 상태인 경기는 제외됩니다.
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: "내 경기 목록 조회 성공"
                data:
                  - matchId: 5
                    title: "백석대 축구 내기"
                    sportType: "SOCCER"
                    matchDate: "2026-05-30T19:00:00"
                    locationName: "백석대 중앙운동장"
                    latitude: 37.6400
                    longitude: 127.0730
                    maxPlayers: 10
                    currentPlayers: 7
                    gender: null
                    ageRange: null
                    status: "OPEN"
                    updatedAt: "2026-05-17T10:00:00"
                    hostId: 1
                    hostName: "홍길동"
                  - matchId: 3
                    title: "농구 3:3"
                    sportType: "BASKETBALL"
                    matchDate: "2026-05-20T18:00:00"
                    locationName: "신부동 공원"
                    latitude: 37.6380
                    longitude: 127.0710
                    maxPlayers: 6
                    currentPlayers: 6
                    gender: "M"
                    ageRange: 20
                    status: "CLOSED"
                    updatedAt: "2026-05-15T09:00:00"
                    hostId: 2
                    hostName: "김철수"
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/members/{memberId}:
    get:
      tags: [회원]
      summary: 다른 회원 프로필 조회
      description: 민감 정보(kakaoId, email, name, phone, role, signupCompleted)는 제외됩니다.
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: "회원 조회 성공"
                data:
                  memberId: 3
                  nickname: "김철수"
                  profileImage: null
                  gender: "M"
                  age: 28
                  address: "수원"
                  skillLevel: "ADVANCED"
                  preferredPosition: "GK"
                  favoriteSports: ["SOCCER"]
        '404':
          description: 존재하지 않는 회원
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example: { success: false, message: "존재하지 않는 유저입니다.", data: null }

  # ──────────────────────────────────────────────
  # 경기
  # ──────────────────────────────────────────────
...
  /api/devices/token:
    post:
      tags: [알림]
      summary: FCM 푸시 토큰 등록/갱신
      description: 앱 실행 시 매번 호출합니다. 같은 토큰이 다른 사용자에게 등록돼 있으면 현재 사용자로 갱신합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceTokenRequest'
            example:
              token: "fGcm_long_token_string..."
              platform: "ANDROID"
      responses:
        '200':
          description: 등록 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: "기기 토큰이 등록되었습니다"
                data: null

    delete:
      tags: [알림]
      summary: FCM 푸시 토큰 삭제
      description: 로그아웃 시 호출합니다. 미호출 시 다른 사용자에게 푸시가 발송될 수 있습니다.
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 삭제 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: "기기 토큰이 삭제되었습니다"
                data: null

  # ──────────────────────────────────────────────
  # 칭찬
  # ──────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "카카오 로그인 후 발급받은 accessToken"
  responses:
    Unauthorized:
      description: 인증 실패 (토큰 없음 또는 만료)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "인증이 필요합니다."
            data: null
...
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          nullable: true
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        data:
          nullable: true
          example: null
    SignUpCompleteRequest:
      type: object
      required: [nickname, gender, age, favoriteSports]
      properties:
        nickname:
          type: string
          minLength: 2
          maxLength: 10
          description: "닉네임 (2~10자)"
        gender:
          type: string
          enum: [M, F]
        age:
          type: integer
        address:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        favoriteSports:
          type: array
          items:
            type: string
            enum: [SOCCER, BASKETBALL, RUNNING, BADMINTON]
          description: "즐겨하는 종목 (1개 이상)"
        skillLevel:
          type: string
          enum: [BEGINNER, INTERMEDIATE, ADVANCED]
        preferredPosition:
          type: string
    UpdateProfileRequest:
      type: object
      description: 변경할 필드만 포함. null이면 해당 필드 유지.
      properties:
        nickname:
          type: string
          minLength: 2
          maxLength: 10
          nullable: true
        address:
          type: string
          nullable: true
        latitude:
          type: number
          format: double
          nullable: true
        longitude:
          type: number
          format: double
          nullable: true
        favoriteSports:
          type: array
          items:
            type: string
            enum: [SOCCER, BASKETBALL, RUNNING, BADMINTON]
          nullable: true
          description: "null이면 유지, 빈 배열이면 전체 삭제"
...
    DeviceTokenRequest:
      type: object
      required: [token, platform]
      properties:
        token:
          type: string
          description: FCM에서 발급받은 푸시 토큰
        platform:
          type: string
          enum: [ANDROID, IOS, WEB]
...